add to basket button created as option for products

+++ website/models.py
@@ -60,14 +60,14 @@ class Category(models.Model):
 
 
 class Product(models.Model):
+
     category = models.ForeignKey(
         'Category', null=True, blank=True, on_delete=models.SET_NULL)
     sku = models.CharField(max_length=254, null=True, blank=True)
     name = models.CharField(max_length=254)
     description = models.TextField()
     price = models.DecimalField(max_digits=6, decimal_places=2)
-    rating = models.DecimalField(
-        max_digits=6, decimal_places=2, null=True, blank=True)
+    rating = models.IntegerField(null=True, blank=True)
     # image_url = models.URLField(max_length=1024, null=True, blank=True)
     # image = models.ImageField(null=True, blank=True)
 
@@ -76,5 +76,6 @@ class Product(models.Model):
 
 
 # quantity = models.IntegerField(max_length=3, null=False)
+# product_id
 
 # orders model
